remove never used migration code from xion module (#353)
I n response to . This pull request removes the migration logic for
transitioning from version 1 to version 2 of the `xion` module. The
changes involve deleting the migration-related code and references
across multiple files.

### Removal of migration logic:

*
[`x/xion/keeper/migrations.go`](diffhunk://#diff-f0a511d6dc13b9840e714db7b1729afd67f07b175d8c064fec7038be30cb6b3bL1-L37):
Deleted the `Migrator` struct, its constructor (`NewMigrator`), and the
`Migrate1to2` method, which handled migration logic based on the chain
ID.
*
[`x/xion/migrations/v1/migration.go`](diffhunk://#diff-156356181f747ac5d4cabef0d6a615ee5d7b268dbdc2d86e51c79f1a5822ef91L1-L98):
Removed the `MigrateStore` function, which implemented the detailed
migration process for updating contracts to a new code ID.

### Cleanup of migration references:

*
[`x/xion/module.go`](diffhunk://#diff-06854a36088167b27c49b6111e70dc1860dc7213868fbd3ce338db68a065dd7bL101-L106):
Removed commented-out code that referenced the `Migrator` and the
`Migrate1to2` method in the `RegisterServices` function, which was
intended for future migration handling.